feat: 내정보수정 파일삭제, 로그인세션 기능 추가
마이페이지 탭 이동 ajax로 수정중

+++ 05_Project/Modu/src/main/java/com/modu/controller/MemberController.java
@@ -1,7 +1,6 @@
 package com.modu.controller;
 
 import java.io.File;
-import java.util.Enumeration;
 import java.util.HashMap;
 
 import javax.servlet.http.HttpServletRequest;
@@ -11,8 +10,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.ModelAndView;
@@ -31,14 +30,12 @@ public class MemberController {
 
 	@Autowired
 	private MemberRegisterService memberRegisterService;
+	
 	@Autowired
 	private FileUploadService filuploadservice; //by @AllArgsConstructor
-	
+		
 	//@Autowired
-	//private FileUtil fileUtil;
-	
-	@Autowired
-	HashMap<String, String> map;
+	//HashMap<String, String> map;
 	
 	//회원가입 페이지 이동
 	@GetMapping("/register")
@@ -50,11 +47,11 @@ public class MemberController {
 	@PostMapping("/register/emailvalidcheck")
 	@ResponseBody
 	public String emailValidCheck(String email){
-		log.info("#1_이메일 중복체크 진입 & email:"+ email);
+		log.info("#1_이메일 중복체크 진입 & 입력한 email:"+ email);
 		int result = memberRegisterService.checkEmail(email);
 		log.info("#2_이메일 result: "+ result);
 		log.info("#3_이메일 길이 email.length(): "+ email.length());
-		if(0<email.length() && email.length()<10) {
+		if(email.length()==0 || 0<email.length() && email.length()<10) {
 			return "noshow";			
 		} else if(result == 0) { 
 			return "success";
@@ -95,7 +92,7 @@ public class MemberController {
 	//회원가입 post - 마케팅 미동의시
 	@PostMapping("/register")
 	public void register(Member member){
-	  log.info("#회원가입 insert 전: Member member= "+ member);
+	  log.info("#회원가입 insert 진입: 파라미터 Member member= "+ member);
 	  memberRegisterService.registerMember(member); 
 	  log.info("#회원가입 성공) Member member= "+ member); //이메일. 닉네임 빼고 모두 비어있음..
 	}
@@ -117,19 +114,20 @@ public class MemberController {
 	//로그인 post
 	@PostMapping("/login")
 	public ModelAndView login(Member member, HttpServletRequest req){
-		log.info("#login 메소드 진입!! 로그인 진입");
+		log.info("#login 로그인 post메소드 진입!!");
 		ModelAndView mv = new ModelAndView();
 		HttpSession session = req.getSession();
 		Member memberInfo = memberRegisterService.login(member); //select EMAIL, NICKNAME from MEMBER where EMAIL=? and PWD=?
 		if(memberInfo == null) {
-			mv.setViewName("member/login");
+			mv.setViewName("member/login"); //이동할 페이지 설정
 			mv.addObject("status", 0);
             return mv; 
 		}else{  
-			mv.setViewName("redirect:/");
+			mv.setViewName("redirect:/"); //이동할 페이지 설정
 			session.setMaxInactiveInterval(1800); //1800초=세션 유효기간 30분으로 지정
 			session.setAttribute("email", memberInfo.getEmail());
 			session.setAttribute("nickname", memberInfo.getNickname());
+			session.setAttribute("profileImg", memberInfo.getProfileImg());
 			return mv; 
 		}
 	} 
@@ -139,34 +137,38 @@ public class MemberController {
 	public String logout(HttpServletRequest req){
 		req.getSession().invalidate();  //세션 무효화
 		req.getSession(true); //새로운 세션 받을 준비 true
-		log.info("#1_(로그아웃 성공 후) HttpServletRequest req: " + req); //org.apache.catalina.connector.RequestFacade@62fcbd5b
+		log.info("#1_(로그아웃 성공!!) HttpServletRequest req: " + req); //org.apache.catalina.connector.RequestFacade@62fcbd5b
 		return "redirect:/";
 	}
 	
 	//마이페이지 페이지 이동
 	@GetMapping("mypage")
-	public String myPage() {
-		return "member/mypage";
+	public ModelAndView goMypage(HttpSession session) {
+		String email = (String)session.getAttribute("email");
+		Member member1 = memberRegisterService.readMyInfo(email); 
+		ModelAndView mv = new ModelAndView("member/mypage", "member", member1); 
+		log.info("######마이페이지 이동get member1: "+member1);
+		log.info("######마이페이지 이동get mv: "+mv);
+		return mv;
 	}
 	
 	//내정보수정 페이지 이동
 	@GetMapping("/modifymyinfo")
 	public ModelAndView goModify(HttpSession session) { 
-		
 		String email = (String)session.getAttribute("email");
-		String nickname = (String)session.getAttribute("nickname");
+		//String nickname = (String)session.getAttribute("nickname");
 		Member member1 = memberRegisterService.readMyInfo(email); 
 		ModelAndView mv = new ModelAndView("member/modifymyinfo", "member", member1); 
 		log.info("######내정보수정 이동get email: "+email);
-		log.info("######내정보수정 이동get nickname: "+nickname);
+		//log.info("######내정보수정 이동get nickname: "+nickname);
 		log.info("######내정보수정 이동get member1: "+member1);
 		log.info("######내정보수정 이동get mv: "+mv);
 		return mv;
 	}
 	
 	//내정보 수정시 닉네임 중복검사 ajax
-	@PostMapping("/modify-myinfo/nicknamevalidcheck")
 	@ResponseBody
+	@PostMapping("/modify-myinfo/nicknamevalidcheck")
 	public String nicknameValidCheck(String nickname, HttpServletRequest req){
 		HttpSession session = req.getSession();
 		int result = memberRegisterService.checkNickname(nickname);
@@ -192,15 +194,14 @@ public class MemberController {
 		
 		log.info("#1_내정보수정 진입  member: "+ member);
 		log.info("#1_내정보수정시 입력 정보 member: "+ member);
-		log.info("#1_현재 세션에 저장된 (수정전) 닉네임 get "+session.getAttribute("nickname")); //스크루바
-		log.info("#1_현재 세션에 저장된 (수정전) 마케팅여부 get "+session.getAttribute("marketing")); //0 = 미동의
+		log.info("#2_현재 세션에 저장된 (수정전) 닉네임 get "+session.getAttribute("nickname")); //스크루바
+		log.info("#2_현재 세션에 저장된 (수정전) 마케팅여부 get "+session.getAttribute("marketing")); //0 = 미동의
 		
 		String ofname = file.getOriginalFilename();
 		log.info("#2_내정보수정 newfile: "+ofname);
 		
 		if(ofname != null)  ofname = ofname.trim();
 		if(ofname.length() == 0) { //파일이 존재하지 않을 때 파일없이 update
-			//memberRegisterService.modifyMyInfo(member);
 			Member memberInfo = memberRegisterService.modifyMyInfo(member);
 			log.info("#3_내정보수정 성공후 내정보 memberInfo= "+ memberInfo);
 			session.setAttribute("nickname", memberInfo.getNickname());
@@ -212,9 +213,9 @@ public class MemberController {
 			log.info("#4_현재 세션에 저장된 (수정후) 프로필사진 get "+session.getAttribute("profileImg"));
 			log.info("#4_현재 세션에 저장된 (수정후) member get "+session.getAttribute("member"));
 			//Member(email=111@naver.com, pwd=1111, nickname=zeze, profileImg=, name=한서인, phone=01055554, marketing=0, apiUsing=0, signupDate=null, updateDate=null, authority=0, point=0)
-			return "redirect:/member/modifymyinfo";
+			return "redirect:mypage";
 			
-		} else {
+		} else { //업로드할 파일이 존재할 때
 			int idx = ofname.lastIndexOf(".");
 			String ofheader = ofname.substring(0,idx); //파일 제목만 뽑기
 			String ext = ofname.substring(idx); //확장자만 뽑기
@@ -227,16 +228,13 @@ public class MemberController {
 			sb.append(ext);
 			String saveFileName = sb.toString();
 			int fsize = (int) file.getSize();	
-			log.info("파일 위");
 			
 			member.setProfileImgOrg(ofname);
 			member.setProfileImg(saveFileName);
 			member.setProfileImgSize(fsize);
-			log.info("파일업로드 위 member:"+ member);
+			log.info("#else문 - 파일업로드 위 member:"+ member);
 			filuploadservice.saveImgFile(file, Path.PROFILE_PATH); //파일업로드서비스단에서 실제 로컬에 물리적 파일 생성
 			
-			//memberRegisterService.modifyMyInfo2(member);
-			log.info("멤버인포 위");
 			Member memberInfo = memberRegisterService.modifyMyInfo2(member);
 			log.info("#5_내정보수정 성공후 내정보 memberInfo= "+ memberInfo);
 			session.setAttribute("nickname", memberInfo.getNickname());
@@ -248,34 +246,41 @@ public class MemberController {
 			log.info("#6_현재 세션에 저장된 (수정후) 프로필사진 get "+session.getAttribute("profileImg"));
 			log.info("#6_현재 세션에 저장된 (수정후) member get "+session.getAttribute("member"));
 			//Member(email=111@naver.com, pwd=1111, nickname=zeze, profileImg=, name=한서인, phone=01055554, marketing=0, apiUsing=0, signupDate=null, updateDate=null, authority=0, point=0)
-			return "redirect:/member/modifymyinfo";
+			return "redirect:mypage";
 		}
-		
 	}  
-		
-	//내정보수정 - 프로필사진 post
-	/*
-	@PostMapping("/modifymyinfo/img")
-	public String updateImg(String email, MultipartHttpServletRequest mpRequest, 
-			HttpSession session) throws Exception  { //HttpServletRequest req 안써도 됨
-		
-		String myImg = fileUtil.updateImg(mpRequest); 
-		Member member1 = (Member) session.getAttribute("login");
-		memberRegisterService.modifyProfileImg(email, myImg);
-		member1.setProfileImg(myImg);
-		session.setAttribute("login", member1);
-				
-		return "redirect:/member/mypage";
-				
-	}                 
-*/
-	//회원 탈퇴         
-	@GetMapping("/remove-myinfo")
-	public String remove(String email, HttpSession session, HttpServletRequest req) { //req 필요
-		memberRegisterService.removeMyInfo(email);
-		session.invalidate(); //현재 접속하고 있는 세션을 무효화
-		//req.getSession(true); //새로운 세션을 받을 준비 true
-		return "redirect:/";
+      
+	//내정보수정 - 현재 나의 프로필사진만 삭제
+	@ResponseBody //주석처리시 return ""의 jsp를 찾아서 리턴
+	@GetMapping("/removemyprofileimg")
+	public String removeMyProfileImg(@RequestParam String profileImg, HttpSession session) { 
+		File file = new File(Path.PROFILE_PATH, profileImg); //PROFILE_PATH 밑에 있는 profileImg을 가져옴 
+		if(file.exists()) {
+			file.delete();
+			String email = (String) session.getAttribute("email");
+			memberRegisterService.removeProfileImg(email); //mapper.xml(sql) 통해 기본사진으로 모두 변경
+			return "delete성공";
+		} else {
+			return "delete실패-파일이 폴더에 존재하지 않음";
+		}
 	}
+	
+	//회원 탈퇴         
+	@PostMapping("/removemyinfo")
+	public String removeMyinfo(@RequestParam("email") String email, HttpSession session, HttpServletRequest req) { //req 필요
+		//System.out.println("#내정보수정 - 회원탈퇴 메소드 진입 email: " + email);
+		//본인이거나 관리자가 아닐경우 삭제 불가하도록
+		Member member = memberRegisterService.readMyInfo(email);	
+		if(member.getEmail().equals((String)session.getAttribute("email"))) {
+			memberRegisterService.removeMyInfo(email);
+			session.invalidate(); //현재 접속하고 있는 세션을 무효화
+			req.getSession(true); //새로운 세션을 받을 준비 true
+			return "redirect:/";
+		} else {
+			return "redirect:/";
+		}
+	}
+	
+
 	
 }
